backend/requirements.txt
falcon
requests
daphne
peewee
numpy
Pillow
joblib
sklearn
requests
python-dotenv
keras
h5py
tensorflow